Move queues to separate files
